+++ ServersideQoL.AutoStore/ItemDropProcessor.cs
@@ -232,6 +232,9 @@ public sealed class ItemDropProcessor : Processor<ItemDropProcessor.PrefabInfo>
           ShowMessage(peers, containerZdo,
               Config.Instance.Localization.Value.FormatAutoPickup(containerState.Container.m_name, item.m_shared.m_name, stack),
               Config.Instance.PickedUpMessageType.Value);
+          var effectPrefab = Instance<ContainerProcessor>().EffectPrefab;
+          if (effectPrefab is not 0)
+            Spawn(effectPrefab, containerZdo.ZDO.GetPosition(), containerZdo.ZDO.GetRotation());
         }
 
         if (item.m_stack is 0)