+++ ServersideQoL.AutoStore/CHANGELOG.md
@@ -1,6 +1,7 @@
 ### v2.0.14
 - Added option `AutoPickupExcludeItemTypes` to exclude certain item types from auto pickup [#232](https://github.com/ArgusMagnus/ValheimServersideQoL/issues/232)
 - Added option `StackInventoryIntoContainersExcludeItemTypes` to exclude certain item types from stacking into nearby containers [#232](https://github.com/ArgusMagnus/ValheimServersideQoL/issues/232)
+- Added option `ShowContainerModifiedEffect` to show an audiovisual effect when a container is modified due to auto pickup or player inventory stacking [#208](https://github.com/ArgusMagnus/ValheimServersideQoL/issues/208)
 
 ### v2.0.11
 - Moved `AutoPickupMaxRange` config option from ContainerSigns to AutoStore